requirements.txt
streamlit>=1.25.0
python-dotenv>=1.0.1
openai>=1.11.0
neo4j>=5.11.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pypdf>=3.13.0
numpy>=1.24.0
pandas>=2.0.0
networkx>=3.0
matplotlib>=3.5.0
